requirements.txt
# Framework Web
Flask==3.0.0
Werkzeug==3.0.1
# Manejo de Excel
openpyxl==3.1.2
# Variables de entorno
python-dotenv==1.0.0
# Servidor de producción
gunicorn==21.2.0
# Utilidades
python-dateutil==2.8.2
# Testing (opcional)
pytest==7.4.3
pytest-flask==1.3.0


# Instalar con:    pip install -r requirements.txt
# Para desarrollo: pip install -r requirements.txt -U